Organize project folders and namespaces
