+++ SemanticKernelChat/Infrastructure/TypeRegistrar.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.DependencyInjection;
 using Spectre.Console.Cli;
 
-namespace SemanticKernelChat;
+namespace SemanticKernelChat.Infrastructure;
 
 public sealed class TypeRegistrar : ITypeRegistrar
 {
